Renamed internal/core to internal/game
