+++ gome.go
@@ -3,7 +3,7 @@ package gome
 import (
 	"github.com/GomeBox/gome/game"
 	"github.com/GomeBox/gome/internal"
-	"github.com/GomeBox/gome/internal/core"
+	internalGame "github.com/GomeBox/gome/internal/game"
 )
 
 //Gome is used to run a game
@@ -16,7 +16,7 @@ type Gome interface {
 //New returns a new instance of Gome
 func New() Gome {
 	g := new(internal.GomeImpl)
-	g.GameRunner = core.NewGameRunner()
+	g.GameRunner = internalGame.NewGameRunner()
 	g.GameSettings = game.NewSettings()
 	return g
 }
