+++ internal/game/contextWrapper.go
@@ -1,4 +1,4 @@
-package core
+package game
 
 import (
 	"github.com/GomeBox/gome/adapters"
